requirements.txt
agent-framework
agent-framework-devui
openai
python-dotenv
uv
streamlit